1.Intern/1.Basic_syntax/7.Arrays/6.5.0.Search_index_in_array/src/test/java/ru/job4j/array/FindLoopWithRangeTest.java
package ru.job4j.array;

import org.junit.Test;

import static org.hamcrest.MatcherAssert.assertThat;
import static org.hamcrest.Matchers.is;

/**
 * Created by Roman Pogorelov on 17.03.2020
 */
public class FindLoopWithRangeTest {
    @Test
    public void whenFind3() {
        int[] input = new int[]{5, 2, 10, 2, 4};
        int value = 2;
        int start = 2;
        int finish = 4;
        int result = FindLoopWithRange.indexOf(input, value, start, finish);
        int expect = 3;
        assertThat(result, is(expect));
    }

    @Test
    public void whenNotFound() {
        int[] input = new int[]{5, 2, 10, 2, 4};
        int value = 5;
        int start = 2;
        int finish = 4;
        int result = FindLoopWithRange.indexOf(input, value, start, finish);
        int expect = -1;
        assertThat(result, is(expect));
    }
}